[maint] make sure that the segmentations are of type integer when loading

+++ refregion/wrappers.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 import nibabel as nib
 from refregion import refregion
+import numpy as np
 
 
 def custom_ref_region(
@@ -15,8 +16,10 @@ def custom_ref_region(
     mask = nib.load(mask_file).get_fdata()
     # create custom reference region
     mask_ref = refregion.custom_ref_region(mask, refregion_indices, erode_by_voxels, exclude_indices, dialate_by_voxels)
+    # cast as uint8
+    mask_ref = mask_ref.astype(np.uint8)
     # save output
-    nib.save(nib.Nifti1Image(mask_ref, nib.load(mask_file).affine), output_file)
+    nib.save(nib.Nifti1Image(mask_ref, nib.load(mask_file).affine), output_file, dtype=np.uint8)
 
 
 def cerebellum_reference_region(
@@ -30,4 +33,8 @@ def cerebellum_reference_region(
     # create cerebellum reference region
     cerebellum_ref = refregion.cerebellum_reference_region(cerebellum, brain)
     # save output
-    nib.save(nib.Nifti1Image(cerebellum_ref, nib.load(cerebellum_segmentation_file).affine), output_reference_region)
+    nib.save(
+        nib.Nifti1Image(cerebellum_ref, nib.load(cerebellum_segmentation_file).affine),
+        output_reference_region,
+        dtype=np.uint8,
+    )
